[TEAM2-64] Refactor : JWT 토큰 에러 처리 리팩토링
