+++ src/main/java/com/rubycon/rubyconteam2/domain/user/exception/UserNotFoundException.java
@@ -1,9 +1,9 @@
 package com.rubycon.rubyconteam2.domain.user.exception;
 
 import com.rubycon.rubyconteam2.global.error.ErrorCode;
-import com.rubycon.rubyconteam2.global.error.exception.EntityNotFoundException;
+import com.rubycon.rubyconteam2.global.error.exception.BusinessException;
 
-public class UserNotFoundException extends EntityNotFoundException {
+public class UserNotFoundException extends BusinessException {
     public UserNotFoundException() {
         super(ErrorCode.USER_NOT_FOUND);
     }
