+++ src/main/java/com/rubycon/rubyconteam2/global/config/oauth/service/CustomOAuth2AuthorizedClientService.java
@@ -1,4 +1,4 @@
-package com.rubycon.rubyconteam2.global.config.oauth.services;
+package com.rubycon.rubyconteam2.global.config.oauth.service;
 
 import com.rubycon.rubyconteam2.domain.user.repository.UserRepository;
 import com.rubycon.rubyconteam2.domain.user.domain.Role;
