+++ src/main/java/com/rubycon/rubyconteam2/global/config/security/SecurityConfig.java
@@ -1,12 +1,12 @@
 package com.rubycon.rubyconteam2.global.config.security;
 
 import com.rubycon.rubyconteam2.global.config.oauth.handler.OAuth2SuccessHandler;
-import com.rubycon.rubyconteam2.global.config.oauth.usertypes.FacebookOAuth2User;
-import com.rubycon.rubyconteam2.global.config.oauth.usertypes.GoogleOAuth2User;
-import com.rubycon.rubyconteam2.global.config.oauth.usertypes.KakaoOAuth2User;
+import com.rubycon.rubyconteam2.global.config.oauth.usertype.FacebookOAuth2User;
+import com.rubycon.rubyconteam2.global.config.oauth.usertype.GoogleOAuth2User;
+import com.rubycon.rubyconteam2.global.config.oauth.usertype.KakaoOAuth2User;
 import com.rubycon.rubyconteam2.global.config.security.entrypoint.JwtAuthenticationEntryPoint;
-import com.rubycon.rubyconteam2.global.config.security.filters.ExceptionHandlerFilter;
-import com.rubycon.rubyconteam2.global.config.security.filters.JwtAuthorizationFilter;
+import com.rubycon.rubyconteam2.global.config.security.filter.ExceptionHandlerFilter;
+import com.rubycon.rubyconteam2.global.config.security.filter.JwtAuthorizationFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
@@ -31,39 +31,39 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
     @Override
     protected void configure(HttpSecurity http) throws Exception {
         http
-                .cors()
+                    .cors()
                 .and()
-                .formLogin().disable()
-                .csrf().disable()
-                .authorizeRequests()
-                .antMatchers(
+                    .formLogin().disable()
+                    .csrf().disable()
+                    .authorizeRequests()
+                    .antMatchers(
                         "/",
                         "/login",
                         "/oauth2/**",
                         "/jwt/**",
                         "/authenticate/**",
                         "/party"
-                ).permitAll()
-                .anyRequest().authenticated()
+                    ).permitAll()
+                    .anyRequest().authenticated()
                 .and()
-                .oauth2Login()
-                // customUserType을 추가하면, 내부적으로 'CustomUserTypesOAuth2UserService' 클래스 사용
-                .userInfoEndpoint()
-                .customUserType(KakaoOAuth2User.class, "kakao")
-                .customUserType(GoogleOAuth2User.class, "google")
-                .customUserType(FacebookOAuth2User.class, "facebook")
-                .and()
-                // 이 부분에서 Success Handler를 설정합니다.
-                .successHandler(oAuth2SuccessHandler)
-                .and()
-                .exceptionHandling().authenticationEntryPoint(unauthorizedHandler);
+                    .oauth2Login()
+                        // customUserType을 추가하면, 내부적으로 'CustomUserTypesOAuth2UserService' 클래스 사용
+                        .userInfoEndpoint()
+                            .customUserType(KakaoOAuth2User.class, "kakao")
+                            .customUserType(GoogleOAuth2User.class, "google")
+                            .customUserType(FacebookOAuth2User.class, "facebook")
+                    .and()
+                        // 이 부분에서 Success Handler를 설정합니다.
+                        .successHandler(oAuth2SuccessHandler)
+                    .and()
+                        .exceptionHandling().authenticationEntryPoint(unauthorizedHandler);
 
         http
-                .sessionManagement()
-                .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+                    .sessionManagement()
+                        .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
                 .and()
-                .addFilterBefore(jwtAuthorizationFilter, UsernamePasswordAuthenticationFilter.class)
-                .addFilterBefore(exceptionHandlerFilter, JwtAuthorizationFilter.class);
+                    .addFilterBefore(jwtAuthorizationFilter, UsernamePasswordAuthenticationFilter.class)
+                    .addFilterBefore(exceptionHandlerFilter, JwtAuthorizationFilter.class);
 
     }
 }
