+++ src/main/java/com/rubycon/rubyconteam2/global/core/jwt/service/JwtService.java
@@ -1,4 +1,4 @@
-package com.rubycon.rubyconteam2.global.core.jwt;
+package com.rubycon.rubyconteam2.global.core.jwt.service;
 
 import io.jsonwebtoken.Claims;
